files/requirements.txt
configparser
paho-mqtt
requests
influxdb